Describe and partly resolve an issue with global functions
Attempting to call certain global functions such as `setTimeout` with
`this` set to the proxy `window` rather than the real window will throw
an error. The Washington Post's GDPR screen for example has code that
gets a reference to `setTimeout` in the global scope and then does
`setTimeout.call(window, ...)` where `window` is the _proxy_ window.