Change approach to overriding `window.location`
The previous approach to overriding `window.location` involved replacing
references to `window` with a proxy which returned a custom value for
`location`. The proxy was implemented with the ES `Proxy` class and this
approach encountered many problems due to proxies being imperfect
substitutes for the target object. In particular attempting to call a
native DOM method with `this` referring to a proxy rather than the real
object (eg. `window.setTimeout(...)`) caused errors in many cases.

The new approach matches what legacy Via does by replacing all
references to the identifier `location` in proxied JS files
with a different name. This approach has limitations in that it will end
up replacing _any_ identifier which happens to have that name, not just
references to the `Location` object. However it actually seems to work
reasonably well in practice because as long as all references to an
identifier are processed in the same way, code continues to work. Also,
and just as importantly, when it does cause problems, it is much easier
to figure out what went wrong.

The initial implementation of this approach is super dumb in that it
just matches "\blocation\b" as a regex, which will also match the word
"location" inside string and regex literals. We can think about making
this more sophisticated later.