Rewrite proxy location implementation
The dummy/fake location object installed as `window.viaLocation` used to
be a JS Proxy that trapped property reads/writes on the real `location`
object.

However since we are intercepting every property there wasn't an
advantage to doing this over creating an ordinary JS class which
implements the `Location` interface, and attempts to call native DOM
methods on the proxy (eg. `toString`) would throw. This caused errors on
sites that tried to treat `window.location` like a string.

This commit rewrites the fake location as a plain JS class instead.